requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
billiard==4.2.0; python_version >= '3.7'
celery[redis]==5.3.6; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
colorama==0.4.6; platform_system == 'Windows'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
fastapi==0.108.0; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.2; python_version >= '3.8'
httpx==0.26.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
kombu==5.3.4; python_version >= '3.8'
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pymongo==4.6.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
redis==5.0.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.24; python_version >= '3.7'
starlette==0.32.0.post1; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.4; python_version >= '2'
uvicorn==0.25.0; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.12
